move admin routes in a seperate folder
