+++ client/src/pages/Admin/EditDetails/style.js
