+++ client/src/pages/index.js
@@ -1,42 +1,9 @@
 import Home from './Home';
 import Step from './Step';
-import CreateUniqueLink from './CreateUniqueLink';
 import Login from './Login';
 import Signup from './Signup';
-import Welcome from './Welcome';
 import ForgotPassword from './ForgotPassword';
 import ResetPassword from './ResetPassword';
-import CreateOrganisationDetails, {
-  SecondStep as CreateOrganisationDetails2,
-} from './CreateOrganisationDetails';
-import StepForm from './StepForm';
-import Dashboard from './Dashboard';
-import Organisations from './Organisations';
-import EditContent from './EditContent';
-import EditLandingPage from './EditLandingPage';
-import ManageSteps from './ManageSteps';
-import EditDetails from './EditDetails';
-import Customise from './Customise';
-import Changes from './Changes';
+import Admin from './Admin';
 
-export {
-  Home,
-  Step,
-  Login,
-  Signup,
-  CreateUniqueLink,
-  Welcome,
-  Dashboard,
-  EditContent,
-  EditLandingPage,
-  CreateOrganisationDetails,
-  CreateOrganisationDetails2,
-  ManageSteps,
-  ForgotPassword,
-  ResetPassword,
-  Organisations,
-  StepForm,
-  EditDetails,
-  Customise,
-  Changes,
-};
+export { Admin, Home, Step, Login, Signup, ForgotPassword, ResetPassword };
